removing some ts errors
