+++ automerge-js/src/bloom.ts
@@ -1,0 +1,124 @@
+/**
+ * Implementation of the data synchronisation protocol that brings a local and a remote document
+ * into the same state. This is typically used when two nodes have been disconnected for some time,
+ * and need to exchange any changes that happened while they were disconnected. The two nodes that
+ * are syncing could be client and server, or server and client, or two peers with symmetric roles.
+ *
+ * The protocol is based on this paper: Martin Kleppmann and Heidi Howard. Byzantine Eventual
+ * Consistency and the Fundamental Limits of Peer-to-Peer Databases. https://arxiv.org/abs/2012.00472
+ *
+ * The protocol assumes that every time a node successfully syncs with another node, it remembers
+ * the current heads (as returned by `Backend.getHeads()`) after the last sync with that node. The
+ * next time we try to sync with the same node, we start from the assumption that the other node's
+ * document version is no older than the outcome of the last sync, so we only need to exchange any
+ * changes that are more recent than the last sync. This assumption may not be true if the other
+ * node did not correctly persist its state (perhaps it crashed before writing the result of the
+ * last sync to disk), and we fall back to sending the entire document in this case.
+ */
+
+import { hexStringToBytes, Encoder, Decoder } from './encoding'
+
+// These constants correspond to a 1% false positive rate. The values can be changed without
+// breaking compatibility of the network protocol, since the parameters used for a particular
+// Bloom filter are encoded in the wire format.
+const BITS_PER_ENTRY = 10, NUM_PROBES = 7
+
+/**
+ * A Bloom filter implementation that can be serialised to a byte array for transmission
+ * over a network. The entries that are added are assumed to already be SHA-256 hashes,
+ * so this implementation does not perform its own hashing.
+ */
+export class BloomFilter {
+  numEntries: number;
+  numBitsPerEntry: number;
+  numProbes: number;
+  bits: Uint8Array;
+
+  constructor (arg) {
+    if (Array.isArray(arg)) {
+      // arg is an array of SHA256 hashes in hexadecimal encoding
+      this.numEntries = arg.length
+      this.numBitsPerEntry = BITS_PER_ENTRY
+      this.numProbes = NUM_PROBES
+      this.bits = new Uint8Array(Math.ceil(this.numEntries * this.numBitsPerEntry / 8))
+      for (const hash of arg) this.addHash(hash)
+    } else if (arg instanceof Uint8Array) {
+      if (arg.byteLength === 0) {
+        this.numEntries = 0
+        this.numBitsPerEntry = 0
+        this.numProbes = 0
+        this.bits = arg
+      } else {
+        const decoder = new Decoder(arg)
+        this.numEntries = decoder.readUint32()
+        this.numBitsPerEntry = decoder.readUint32()
+        this.numProbes = decoder.readUint32()
+        this.bits = decoder.readRawBytes(Math.ceil(this.numEntries * this.numBitsPerEntry / 8))
+      }
+    } else {
+      throw new TypeError('invalid argument')
+    }
+  }
+
+  /**
+   * Returns the Bloom filter state, encoded as a byte array.
+   */
+  get bytes() {
+    if (this.numEntries === 0) return new Uint8Array(0)
+    const encoder = new Encoder()
+    encoder.appendUint32(this.numEntries)
+    encoder.appendUint32(this.numBitsPerEntry)
+    encoder.appendUint32(this.numProbes)
+    encoder.appendRawBytes(this.bits)
+    return encoder.buffer
+  }
+
+  /**
+   * Given a SHA-256 hash (as hex string), returns an array of probe indexes indicating which bits
+   * in the Bloom filter need to be tested or set for this particular entry. We do this by
+   * interpreting the first 12 bytes of the hash as three little-endian 32-bit unsigned integers,
+   * and then using triple hashing to compute the probe indexes. The algorithm comes from:
+   *
+   * Peter C. Dillinger and Panagiotis Manolios. Bloom Filters in Probabilistic Verification.
+   * 5th International Conference on Formal Methods in Computer-Aided Design (FMCAD), November 2004.
+   * http://www.ccis.northeastern.edu/home/pete/pub/bloom-filters-verification.pdf
+   */
+  getProbes(hash) {
+    const hashBytes = hexStringToBytes(hash), modulo = 8 * this.bits.byteLength
+    if (hashBytes.byteLength !== 32) throw new RangeError(`Not a 256-bit hash: ${hash}`)
+    // on the next three lines, the right shift means interpret value as unsigned
+    let x = ((hashBytes[0] | hashBytes[1] << 8 | hashBytes[2]  << 16 | hashBytes[3]  << 24) >>> 0) % modulo
+    let y = ((hashBytes[4] | hashBytes[5] << 8 | hashBytes[6]  << 16 | hashBytes[7]  << 24) >>> 0) % modulo
+    const z = ((hashBytes[8] | hashBytes[9] << 8 | hashBytes[10] << 16 | hashBytes[11] << 24) >>> 0) % modulo
+    const probes = [x]
+    for (let i = 1; i < this.numProbes; i++) {
+      x = (x + y) % modulo
+      y = (y + z) % modulo
+      probes.push(x)
+    }
+    return probes
+  }
+
+  /**
+   * Sets the Bloom filter bits corresponding to a given SHA-256 hash (given as hex string).
+   */
+  addHash(hash) {
+    for (const probe of this.getProbes(hash)) {
+      this.bits[probe >>> 3] |= 1 << (probe & 7)
+    }
+  }
+
+  /**
+   * Tests whether a given SHA-256 hash (given as hex string) is contained in the Bloom filter.
+   */
+  containsHash(hash) {
+    if (this.numEntries === 0) return false
+    for (const probe of this.getProbes(hash)) {
+      if ((this.bits[probe >>> 3] & (1 << (probe & 7))) === 0) {
+        return false
+      }
+    }
+    return true
+  }
+}
+
