+++ automerge-js/test/sync_test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert'
 import * as Automerge from '../src'
-import { BloomFilter } from '../src/sync'
+import { BloomFilter } from '../src/bloom'
 import { decodeChangeMeta } from '../src/columnar'
 import { decodeSyncMessage, encodeSyncMessage, decodeSyncState, encodeSyncState, initSyncState } from "../src"
 import * as AutomergeWASM from "automerge-wasm"
